Error fixing
BabotsWifiAnalyzer.c:43:18: error: assigning to 'int' from incompatible type 'void'
    num_networks = scan_networks();
                 ^ ~~~~~~~~~~~~~~~
BabotsWifiAnalyzer.c:58:5: error: implicit declaration of function 'get_ssid' is invalid in C99 [-Werror,-Wimplicit-function-declaration]
    get_ssid(ssid);
    ^
BabotsWifiAnalyzer.c:61:24: error: implicit declaration of function 'get_signal_strength' is invalid in C99 [-Werror,-Wimplicit-function-declaration]
    *signal_strength = get_signal_strength();
                       ^
BabotsWifiAnalyzer.c:64:22: error: implicit declaration of function 'get_channel_usage' is invalid in C99 [-Werror,-Wimplicit-function-declaration]
    *channel_usage = get_channel_usage();
                     ^
4 errors generated.

Explain the errors

The first error is telling us that we are assigning a value of type 'void' to a variable of type 'int', which is not allowed. This is occurring in line 43 of the program, when we are trying to assign the return value of the function 'scan_networks' to the variable 'num_networks'.

The second error is telling us that the function 'get_ssid' has not been declared, which is causing an implicit declaration error. This is occurring in line 58 of the program, when we are trying to use the function 'get_ssid' to get the SSID of the network.

The third error is telling us that the function 'get_signal_strength' has not been declared, which is causing an implicit declaration error. This is occurring in line 61 of the program, when we are trying to use the function 'get_signal_strength' to get the signal strength of the network.

The fourth error is telling us that the function 'get_channel_usage' has not been declared, which is causing an implicit declaration error. This is occurring in line 64 of the program, when we are trying to use the function 'get_channel_usage' to get the channel usage of the network.

hwo do we fix it?

To fix these errors, we need to declare the functions 'get_ssid', 'get_signal_strength', and 'get_channel_usage' at the beginning of the program. We can then implement the functions and use them in our program.