He keeps using the name
This program takes a MAC address as an argument and uses the wifi_geolocation_lookup() function to look up the device's location. It then prints out the latitude and longitude of the device. This program is able to quickly and accurately track the location of a device using its MAC address.